+++ src/rules/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Rules for @beanbag/eslint-plugin.
+ */
+
+import spinaWrappedStaticAccess from './spina-wrapped-static-access';
+
+
+const rules = {
+    'spina-wrapped-static-access': spinaWrappedStaticAccess,
+};
+
+export default rules;